Workaround for DeploymentUnit loading
The Stream/Job factories require a deployment path such as
/xd/deployments/streams/<name> to write out the deployment
properties. This path is read by the factories to create
a new instance of DeploymentUnit. This commit has a "workaround"
to place those properties in the path before the DU is loaded.

This is required because ResourceDeployer has been modified
to work directly with DU instances instead of String IDs.